# Optimize `TimeSeriesSlice` performance