*Description of changes:* the `TimeSeriesSlice` class used in `gluonts.dataset.split` tools adds a lot of overhead due to pandas conversions. This PR gets rid of most of the overhead by operating directly on `numpy` arrays.

To test the improvement one can use the following snippet, iterating `N` times over the `m4_hourly` dataset and splits obtained from it:

```python
import time
from contextlib import AbstractContextManager

from gluonts.dataset.repository.datasets import get_dataset
from gluonts.dataset.split import split

dataset = get_dataset("m4_hourly").train
# dataset = list(get_dataset("m4_hourly").train)

class Timer(AbstractContextManager):
    def __init__(self, name: str):
        self.name = name

    def __enter__(self):
        self.start = time.time()

    def __exit__(self, *args):
        end = time.time()
        print(f"{self.name}: {end - self.start} seconds")

N = 100

with Timer("dataset    "):
    for k in range(N):
        for entry in dataset:
            pass

training_data, test_template = split(dataset, offset=-100)

with Timer("training   "):
    for k in range(N):
        for entry in training_data:
            pass

test_pairs = test_template.generate_instances(prediction_length=24)

with Timer("test pairs "):
    for k in range(N):
        for entry in test_pairs:
            pass
```

Before the fix:

```
dataset    : 3.6620118618011475 seconds
training   : 13.0590181350708 seconds
test pairs : 22.129828214645386 seconds
```

After the fix:

```
dataset    : 3.6478400230407715 seconds
training   : 3.8759350776672363 seconds
test pairs : 4.291091203689575 seconds
```

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.


**Please tag this pr with at least one of these labels to make our release process faster:** BREAKING, new feature, bug fix, other change, dev setup